+++ packages/solidart/lib/src/core/core.dart
@@ -4,7 +4,8 @@ import 'dart:convert';
 import 'dart:developer' as dev;
 import 'dart:math';
 
-import 'package:alien_signals/alien_signals.dart' as alien;
+import 'package:alien_signals/preset.dart' as alien;
+import 'package:alien_signals/system.dart' as alien_system;
 import 'package:collection/collection.dart';
 import 'package:meta/meta.dart';
 import 'package:solidart/src/extensions/until.dart';